requirements.txt
streamlit
cryptography
google-cloud
google-auth
requests
google-cloud-aiplatform
